Repo: Create hooks Filters and Requests
